Enable autotools build support for kinetic-cpp-client
The reasons for moving to autotools build infrastrcture are following:
* Seamlessly enable cross builds.
* Seamlessly enable various architecture builds.
* Use shared libraries instead of building and embedding all the external
  code in one binary.

The autotools build is documented in the README.md file.
In support of the autotools build infrastructure the older cmake build
support is removed. All the tarballs of external sources are dropped, as
now the build is using external libraries to get the same support.

The travis build configuration is also updated to work with autotools build

Signed-off-by: Kinjal Patel <kcpatel83@yahoo.com>
Signed-off-by: Nitin A Kamble <nitinkamble@yahoo.com>